requirements.txt
pyTelegramBotAPI==3.6.7
matplotlib==3.1.3
python-dotenv==0.10.5